Add categories, description, and screenshots to Application panel
This CL makes sure the manifest fields below are shown in the
Applications panel:
- categories https://w3c.github.io/manifest-app-info/#categories-member
- description https://w3c.github.io/manifest-app-info//#description-member
- screenshots https://w3c.github.io/manifest-app-info/#screenshots-member

It also adds some basic validation errors for screenshots, such as a min
size of 320×320.

Those are shown currently in the experimental PWA install UI using a
bottom sheet on Android.

Screenshots: https://imgur.com/a/eeYv706
Bug: 1146450
Change-Id: I3b9da7615332d63c93ca1f8e4b12f78545fb0f60