Skip GRD compression in debug builds
GRD compression takes a while. It takes a particularly long while in
debug builds. This change skips compression on debug builds which
takes the GRD step from ~55 seconds to ~5 seconds.

Alternately we could use this framework (passing through the debug state
of the build) to try different compression options as suggested in the
bug.

Bug: 1162467
Change-Id: I017ec872a1645fef55aff09472cdf4427146d17b