Avoid rendering hidden DOM in data-grid and batch scrolls
This CL lands some low hanging perf improvements to the data-grid, which
slows when large amounts of rows are created in quick succession (see
attached bug).

This CL does not solve the performance issues noted in the bug but does
improve them a fair amount based on my testing. The data-grid was
creating DOM for columns that were entirely hidden, so we don't do that
anymore. We can also defer the scrollToBottom logic into a
requestAnimationFrame to stop the browser having to layout so often.

Bug: chromium:1169021
Change-Id: Iea2f4a2a5cc12c7b15517fa163eec1c6cd787cb9