Show document resource for iframes written to with 'document.write()'
This CL takes care of rendering a placeholder for the document resource
of an iframe which has been written to via JS's 'document.write()'.
The backend considers the document containing the 'document.write()'
call as the document of the frame being written to. Since this could
confuse users, instead of showing this document or showing no document
at all, this CL renders a placeholder "Document not available" resource,
which explains the situation to the user.

Bug: 1107766

Change-Id: Ie982b85ae89ccc8d8a95790fd1b9e002639859fb