Fix SW view background unexpected shown in a narrow pane
If the application pane is narrow, a service workers update timeline may
exceed the boundary of the boundary of the table it is in, causing the
background of Application pane to be shown.

Also, an accessibility issue is also fixed: when keyboard moves to
a timeline row, one couldn't open the details using keyboard. “Switch"
role was set on the wrong element.

Bug: 1169596
Bug: 1169663
Change-Id: I7583bd763c7387cc56f59f1af2d30653eb589c9a