Add warning for manifest description length to Application panel
This CL makes sure that the user gets a warning in Application panel if
the description of the manifest won't be featured in the experimental
PWA install UI using a bottom sheet on Android. It happens if its length
is less than 80 characters and more than 324 characters.

Chromium CL: https://chromium-review.googlesource.com/c/chromium/src/+/2642690
No-Presubmit: true
Bug: 1146450
Change-Id: Ie6a79a6f44dd5ffa2e5842aa8002b49e1a2b1447