[module.json extensions] Group views by location before sorting
Some views define an 'order' field so that views that share a location
can be sorted in a predefined order. Because some views don't define
such field, thus, sorting the complete set of registered extensions by
'order' doesn't necessarily produce an array in which views within a
location are in order with respect to one another.

As a workaround a grouping by location is introduced before sorting
registered views in ViewManager.

Bug: 1131500
Fixed: 1170241
Change-Id: Iaf270c73e160c94fce5e0ed52e1c3f2473250990