Settings: Improve display of shortcuts on Mac and non-US keyboards
Currently, the shortcut editor mostly just uses the KeyboardEvent.key
field to determine what to display in response to user input. This works
well for en-us keyboard on Windows, but often results in confusing
behavior on keyboards which input special characters with the alt key,
such as Option+` e to produce è on the default Mac US keyboard. In order
to better represent these keypresses, this CL adds KeyboardEvent.code as
a fallback for determining how to display a key in a shortcut.
The W3C details the relationship between physical keys and keycodes at
[1], but it is essentially a string equivalent to the now-deprecated
keyCode field that the DevTools have long used to determine whether
a keypress matches a shortcut. Some keyboards with different layouts
(e.g. AZERTY) will display their QWERTY equivalents--pressing Ctrl+A
would display CTRL+Q--but this reflects the underlying implementation,
since shortcuts are saved by their keycodes.

[1] https://www.w3.org/TR/uievents-code/#keyboard-key-codes

Bug: 174309
Change-Id: Ieff84821641d3206ec67b05658f6b23f9d0d36e4