[lmi] Speed up data retrieval and cleanup tests.
This introduces a new RemoteArrayBuffer abstraction similar to the
RemoteArray abstraction in SDK.RemoteObject, and uses that for the
linear memory inspector, which now no longer needs the artificial
Uint8Array all the time. Whenever it needs a bunch of bytes, it'll
now retrieve them with a single CDP `Runtime.callFunctionOn` instead
of having one call per byte.

Bug: chromium:1110202
Also-By: kimanh@chromium.org
Change-Id: I218d7573f58de971f3166704e8c93e5b79533bc0