Introduce GN node_action template
This template abstracts away from the `node.py` script, as well as
making sure that the JS script is properly added to the inputs
array. Before, it was easy to forget to add the JS script to the
inputs array, which can lead to GN incorrectly determining there
is no need to rebuild, while in fact the script itself changed.

R=szuend@chromium.org

Bug: 1170706
Change-Id: I759f3ef3088381ea7e4cefcb8524d7d240b46fb9