Render a subset of rows in the data-grid
This CL drastically increases the performance of the data-grid on large datasets by only rendering the rows that are visible to the user. Any offscreen rows do not get rendered at all, so even if the table has 50000 rows, we only render (as an example) the 20 that the user can see, plus a few more for padding on either side.

Change-Id: I92bd464dbb958f51488337a0da3149ba9a399653
Bug: chromium:1169021
Also-By: aerotwist@chromium.org