Fix squashed profile text in Throttling settings
By upping the flex-basis to 100px, that meant that the other inputs
didn't size down from 100px, meaning that there was no room left for the
profile name box to stay a reasonable size. Bringing that 100 down to 70
ensures there is room to show the profile name too.

See https://imgur.com/a/yQFKF8Z for screenshot of fix.

Fixed: chromium:1170314
Change-Id: Ic78b8527e668782bb0feff53cbd900867f75636f