[M88] Fix repaint bug for chrome remote debugging
Manually cherry-picked from clrev.com/c/2652510 due to merge conflicts.

This CL fixes a bug introduced in https://crrev.com/c/2445770. The
design of highlightInOverlay is misleading, as it should be called
updateHighlightInOverlayAndRepaint (because it is also responsible
for resetting the highlight, and, in the absence of a highlighted
node, for repainting the image). Another reminder that good method
names matter.

I'm deferring renaming the method to a later CL so we can keep this
small and back-mergeable.

Bug: chromium:1158340
Change-Id: Ic94d63eef8aa60e5e727514f51e2ca22984f2257