Fix dimensions of the hidden color input in the Layout panel
The 'Show a quick highlight on the focused object' feature
available in chrome://settings/accessibility draws a highlight
around focused objects. The color input in the Layout panel
is not visible and didn't have proper dimensions although the
a11y still used its dimensions to draw the highlight. This CL
fixes the dimension and the position of the hidden color input.

Fixed: chromium:1171261
Change-Id: Ib57673f5bff74a2da47ca10f6d369eb1888f3851