[l10n] Add a presubmit that runs the bundling step for the i18n lib
Previously we had no check that this bundling step was run when
the i18n lib source file was changed. Now presubmit will run the
bundling if the relevant devtools files change.

We don't check for changes to rollup itself or any of the other
non-devtools deps of the i18n lib at this point even though they
would change the output.

Bug: 941561
Change-Id: I9436f62b45f5aa4a2f2fc339941113f11962b7d0