Use build config from build/
When we moved to a separate repository, we initially used the
`build_overrides/` folder to specify our `BUILDCONFIG.gn`. However, we
should instead use the BUILDCONFIG from `build/config`, which is
maintained in the `build` repository.

This change is necessary to fix an issue where the order of loading the
`BUILDCONFIG.gn` could lead to several variables (most notably the
platform variables such as `is_linux`) to be incorrect. That happened
in the situation where the `current_os` was incorrectly kept empty,
because we weren't setting the proper toolchain arguments. All of that
is correctly implemented in `build/config/` and therefore it is better
to use these.

R=liviurau@chromium.org

Bug: 1139220
Change-Id: Iea7b44dc4b991540bc75881953cd723f8487f75c