Migrates DataGrid over to coordinated render
This CL migrates all the various rAF calls in the DataGrid over to using
the read/write/scroll handlers in the RenderCoordinator. It also
includes a drive-by fix for the RenderCoordinator itself: a convenience
function for being made aware when the current workload is done, and a
minor ordering fix to ensure we don't declare the work done too early.

Bug: 1171202
Change-Id: I078c90aa46ca1b83c52395895cfd155055db1341