Add warnings for manifest screenshots to Application panel
This CL makes sure that the user gets a warning in Application panel if
the screenshots max dimensions are twice larger than the min dimension.
If so, those won't be featured in the experimental PWA install UI using
a bottom sheet on Android.

Chromium CL: https://chromium-review.googlesource.com/c/chromium/src/+/2659198

Bug: 1146450
Change-Id: Ie017d8c1d7e6715bd26ea3d339d292af380df43c