Add support for labelling render coordinator actions
Render coordination can be challenging to debug, since work is
inherently spread across multiple frames. This CL adds logging support
to the Coordinator, and as a drive-by, it adds labels to the key parts
of the DataGrid's coordination work. (Side note: this allowed me to
find and remove a couple of unnecessary awaits / coordinator calls.)

Bug: 1171202
Change-Id: I68cf1ca95cb6d762eaf0f080c8393b551fdfe9da