Introduce flag to hosted server to disable remapping
The remapping behaviour means if I run `npm run server` in my checkout
and hit, as an example,
https://localhost:8090/front_end/theme_support/theme_support_impl.js, I
get a 404, because it's been remapped to ../resources/inspector/... and
that isn't where my Chromium checkout happens to be.

Therefore introducing this flag to disable the remapping such that you
can run the server and have files served up locally.

Bug: chromium:1164945
Change-Id: Ia7c9c23535c88f45125a69457f01d175eb6291b7