[wasm] Conditional breakpoints, logpoints and 'Never pause here'.
Following up on https://crrev.com/c/2666690, this enables support for
the following features in DevTools front-end:

 - Conditional breakpoints
 - Logpoints
 - Never pause here

All of these are based on conditional breakpoints support in V8, which
are now supported in V8 for Wasm. The condition is always a JavaScript
snippet that is executed via "Local Debug-Evaluate", and hence has
access to the state of the Wasm engine via the Debug Proxy API.

Note that for sources provided via language plugins we keep the relevant
context menus hidden, since it's not yet clear how we'd go about
supporting expression evaluation there (i.e. we will need some kind of
basic C++ expression support in case of the DWARF extension).

Fixed: chromium:1173006
Bug: chromium:1164241, chromium:1071432, chromium:1142275
Change-Id: Ib26875f38f1f26388d3ee772e95424fdc17e5eb9