Preemptively check locations to handle undefined or invokable titles
The `emulation.locations` setting represents the locations visible in
the location settings tab. When the setting value is updated, it is
serialized and cached.

crrev.com/c/2650505 wraps the `title` field of this setting's default
values in functions so that the titles can be localized, while
updating the references accordingly. Nevertheless a preemptive check
is necessary as the values are serialized before a reference is made
to the `title` field.

This CL implements such preemptive check so that the field remains
serializable when the setting's value is cached.

Also, because users affected by the bug would have cached undefined
titles, another correction is implemented so that any default value
mistakenly stored without a title is replaced with the corresponding
declared value of the pre-registered setting.

Fixed: 1171670, 1172220, 1172214
Change-Id: Ia403fbba1cca19b022cfb2b7d1b6f8dad768a889