Enforce camelCase for private and protected class members
This encodes some of the TypeScript style guides about identifiers
https://google.github.io/styleguide/tsguide.html#identifiers
(in particular the rules our codebase already adheres to).
The automatic enforcement will save time during code reviews.

The current TypeScript migration produces a lot of public class
properties and methods with leading underscores, so this CL only
enforces the rule for protected and private class members; the
rule can be enforced for public variables as well at a later
point in time.

Bug: chromium:1057042
Change-Id: Ic63b0a9b128f8b560020c015a0f91bbf72a9dd6d