[lmi] Add memory icon to Scope and Console preview.
This introduces a new memory icon next to ArrayBuffer, DataView,
TypedArray, and WebAssembly.Memory instances in the Scope and Console
previews which opens the linear memory inspector.

The current approach is a bit hacky, and should probably be revisted as
we redesign the object UI, but for now it's sufficient to make the
necessary progress here.

Screenshot: https://imgur.com/o5pSDGg.png
Fixed: chromium:1174063
Bug: chromium:1166577
Change-Id: I74b7f09e59c40c35cc3ac07c99bee31473224f79