[module.json extensions] Migrate unit tests out form legacy extensions infra
Currently, the environment helpers for unit tests instantiate a Runtime
singleton and provide the settings necessary to bootup when creating
this instance. As part of the effort to migrate away from module.json
extensions we need to register these settings using the new JavaScript
based system in EnvironmenHelpers.ts.

For this purpose, an auxiliary method is implemented and used to
register the settings needed by untit tests. The method also allows to
reset the list of currently registered settings if required.

The call to conditionaly instantiate a Runtime singleton is not removed
from the environment helpers as there are still several dependencies on
a Runtime instance by the tests.

Bug: 1134103
Change-Id: Ib5cc19a1abcf004b1fdfbbdd29b5e81f07f6a24c