[lmi] Reduce visual clutter on TypedArrays.
The previously added memory icon adds too much visual clutter to
TypedArrays which already come with an element preview. So this CL
restricts the memory icon to just ArrayBuffer and WebAssembly.Memory
for now, and leaves DataViews and TypedArrays unchanged. For the latter
it's easy to get to their buffers and click on the memory icon of the
buffer.

Bug: chromium:1166577, chromium:1174063
Change-Id: I316da8d079690f7917dfb0e4c71e69325a2f7f05