[l10n] Check in en-XL.json
We use this debug locale file in e2e tests now, meaning the bots need
this generated file to run tests. Check it in, similar to en-US.json.

Update the presubmit warning to also include en-XL.json.

Presubmit already runs collect-strings.js which generates this file,
which is no longer git-ignored. We will now get presubmit errors if
changes are not checked in, just like with en-US.json.

No-Presubmit: true
Bug: 1136655
Change-Id: I6ba790960fe722a0b453f6687345b41e5e97e8b6