[TreeOutline] Initial component structure.
IMPORTANT: this component is *not* ready for production use.

Rather than build the entire TreeOutline and land it in one huge CL, I
am instead landing it in small pieces. This CL includes:

* the basic TreeOutline component
* full keyboard navigation [following the WAI-ARIA guidelines:
  https://www.w3.org/TR/wai-aria-practices/examples/treeview/treeview-2/treeview-2b.html]
* support for recursively expanding and collapsing from nodes.
* thorough unit tests and component doc examples.

It does not include work on being able to customise rendering of nodes
(that's next), or CSS work, so it doesn't look like it is meant to
currently.

Bug: chromium:1166669
Change-Id: I4a3ce6e0553453067d05359ae9ef3150081250d7