Remove front_end/toolbox.json
Since all of its modules are removed, we can now remove toolbox
as an entrypoint from build_release_applications and build it
with `devtools_entrypoint`. There was one catch: the copy logic
we use as part of `devtools_entrypoint` was wrong and required
us to remove the trailing slash, since `front_end/toolbox.js` isn't
in a subdirectory.

R=andoli@chromium.org,aerotwist@chromium.org

Bug: 1127902
Change-Id: Ie8b47a5ceae66c8e6b607794d564d17dc6d1e3af