Patch Mocha's describe to include test names
When tests fail, debugging the failure includes isolating the test before
opening it in debug mode or adding diagnostics to it. With the current
test output, opening the file is surprisingly tedious, since there is no
direct way to map test log output to the tests on the filesystem. One has
to grep the test directory for test names first. This CL wraps Mocha's
describe function to automatically include the test filename into the
test title in a format that can be copy-pasted into the test runner's
--test-file argument.

Bug: none
Change-Id: I172bd14f5026ec0041b933c64bcdcccf58536fb9