Stop copying to resources/inspector in devtools_{module,entrypoint}
This simplifies the JavaScript generating tasks in the DevTools
build system to stop copying to resources/inspector. The output
to the gen-folder will remain as-is.

After this change, it will no longer be possible to use
resources/inspector as build output location when using
--custom-devtools-frontend. Instead, the location should be
updated to use gen/front_end instead. Engineers are encouraged
to remove the resources/inspector build output folder to ensure
stale build artifacts aren't accidentally used while working
on DevTools locally.

Note that there are still files in resources/inspector generated by
other actions in the build system. We will clean these up in
follow-up Cls.

R=alexrudenko@chromium.org,jacktfranklin@chromium.org

Bug: 1174013
Change-Id: I750d04e58c5fa85054b8451c3582b58a3f7c77ba