Suggestion: put UIStrings at the end of the file
This CL updates the documentation to suggest adding the
UIStrings to the end of the file. The reason is that
UIStrings can be quite long and to get to the main content,
one might need to scroll down more. Unless there are technical
reasons to keep UIStrings at the top, I'd suggest putting them
at the end of the file.

Bug: chromium:1176653
Change-Id: Id6e6b9291d69324b88310f00f89d3d7626aaeafe