Inject custom scrollbar CSS into components
The new components weren't getting the scrollbar styles because of
ShadowDOM, so I've split them into a file that can then be adopted from
within a component.

See bug for before, and https://imgur.com/a/ba7W4xL for after. Checked
also that this still works in scrollbars not in a ShadowDOM.

Fixed: chromium: 1176554
Change-Id: I823ed80335c73ae681b172823ea72053e2d3a2ae