Add legacy module loader to TestRunner
For tests that import certain modules, we have to import the correct
legacy modules. The `module.json` historically fulfilled these tasks,
but we are removing these files. Therefore, we need another method to
import the legacy file dynamically in layout tests.

We don't want to statically import all legacy files in the
integration_test_runner.js, as that will load too much code and could
let the tests time out. Therefore, we need to add another method to
solely import the legacy file, but nothing from the Runtime machinery.

R=aerotwist@chromium.org

Bug: 1127902
Change-Id: I433692bab9ac69a39a1cc224695722b373db9f12