Add check for platform to set settings' titles on bootup
When setting up the infrastructure to migrate extensions of type
'setting' into a JS based system, a check for the existence of the
`titleMac` field was added to determine if a setting has a different
title in case DevTools is ran on a mac. However, a check for the current
platform was missing before determining the setting's title. This made
the titleMac field always used if present regardless of the platform
running DevTools.

This adds such check so that the titleMac field is used only when the
running platform is mac.

Fixed: 1176719
Change-Id: I27972536d52bb97bd06e7f86cef9c0e6c8a31769