[Flexbox] Overlay front-end for drawing flex items
If the backend responds back with the information for a flex item, then
this draws the base size and flexibility arrow, like so:

+---+-------+
|///|       |
|///|------>|
|///|       |
+---+-------+

The idea is when there is a base size we can detect (which can come from
flex-basis or width or height), then we draw this base size area, and
also draw an arrow from that base size to the final size of the item.

The base size can be smaller than the item, or bigger, if the item
shrank.

+-----------+ -  -  - +
|///////////|/////////|
|///////////|<--------|
|///////////|/////////|
+-----------+ -  -  - +

Screenshot of the new debug page: https://imgur.com/qqjsUwW.jpg

Bug: 1173995
Change-Id: I8e16785c58b801f09340317eae0b922ab1c59422