Only lint changed CSS files
Previously, we'd always lint all CSS files, even the unchanged ones,
because the --files switch was missing in the call to
run_lint_check_css.js, so it would fall back to the **/*.css glob.

Another drive-by is that it corrects two typos, one of them introduced
in my last CL, the other preventing a full lint in case a stylelint
source file or configuration file was changed.

Bug: chromium:1166108
Change-Id: Ia574d77a0fcacddc82238fe5b51a02b375aaa457