Enforce camelCase (with optional leading underscore) for public methods
The leading underscore is accepted for the time being because the
automated JSDOC -> TS translation produces such identifiers, and
renaming will take some time.

We're not enfocing any convention on object literals, as they are used
as arguments to library functions in some places, which might follow
different styles.

This CL also enforces PascalCase for interface, typeAlias, and
typeParameter.

Bug: chromium:1057042
Change-Id: I8decfae35a2231d8068ce50b230cb1866ec494da