Show sourcePort and sourceScheme in CookiesTable
This CL also introduces basic editing for the values, and makes the
SameParty attribute editable as well.

Corresponding back-end CL: https://crrev.com/c/2584499

The current UX is that once a cookie is edited, and the back-end
reports an error setting the cookie, the line in the cookies table
is marked red. The user can then continue to edit the cookie until
it is accepted by the back-end. Unfortunately, the imprecise error
reporting makes it impossible to give a good explanation as to why
the cookie failed to set; this is future work, we want to pipe the
cookie inclusion status to the front-end (which can then use similar
logic as the network panel to explain why setting failed).

DISABLE_THIRD_PARTY_CHECK=protocol update pending

Bug: chromium:1170548, chromium:1142606
Change-Id: Ie13894942318f78311c3c0c3d4a4491819c10647