Remove front_end/client-variations
All usages of client-variations now directly import the third_party
content. This was previously not possible with the module.json
infrastructure, but we can now directly import the relevant
files. Therefore, we can remove the module.json file itself and
the references to it in the module.json dependency hierarchy.

R=mathias@chromium.org

Bug: 1127902
Change-Id: I36d987ca3e645407879da0a81f305de5088711e5