[Styles - Accessibility]: Add SR support for invalid property name and values
Issue:
- Warnings for invalid properties in the styles sidebar pane are not SR accessible
    - example: "Invalid property value" and "Invalid property name"

Changes:
- Added an alert when creating the exclamation triangle next to invalid property names

Comments:
This change does not produce the ideal behavior for screen readers, but this is the least invasive solution to exposing the warning to screen readers.
The main issue stems from how the styles properties and  are created and repopulated.  Since this implementation will announce the warning whenever the warning icon is created, the warning will be announced every time the property is re-populated.

Properties are repopulated when:
- that property's blur handler is triggered on either the name or value inputs
- when any property in that section is modified
- when adding or removing any property in that section


Bug: 1166435
Change-Id: I960b495735f211c5d8404a3b1083172c1911a7e6